achordion + keymap: Support for achordion
Support achordion, a library to help with home row
modifier usage, and bottom row modifier usage.

This also adds the ability to hold down two layers at once, in
3 pairs, 23, 45 and if both are held 67.

Along with this, there is a change to allow KC_TRANSPARENT
to be used along with mod tap, as long as the mod_tap is on
the top active layer.  This is needed for bottom row modifier
support on Dvorak to have a sane NAS layer.