Update to code as merged in mainline QMK.
Chordal Hold has now merged in the mainline QMK repo develop branch.
This commit updates the PR to the exact version of the core code when it
merged, to avoid conflict in the future.

I've also revised to remove the Chordal Hold enable boolean from the
core code, and implemented a functionally equivalent enable/disable
toggle at the keyboard level.